Add delete confirm message box
v2->v3 Rebase the code
v1->v2 Change some message translations in po files.
       Make the confirm button text and cancel button
       text parameters, so that the message box can be
       used more widely.
1 Add a delete confirm message box when
delete template and vm.
issue:https://github.com/kimchi-project/kimchi/issues/85

+++ ui/js/src/kimchi.template_main.js
@@ -54,9 +54,18 @@ kimchi.bindClick=function() {
 		kimchi.window.open("template-edit.html");
 	});
 	$('.template-delete').on('click', function(event) {
-		var templateName = $(this).data('template');
-		kimchi.deleteTemplate(templateName,"","");
-		kimchi.doListTemplates();
+        var $template = $(this);
+        var settings = {
+                title: i18n['msg.confirm.delete.title'],
+                content: i18n['msg.template.confirm.delete'],
+                confirm: i18n['msg.confirm.delete.confirm'],
+                cancel: i18n['msg.confirm.delete.cancel']
+            };
+        kimchi.confirm(settings,function() {
+            var templateName = $template.data('template');
+            kimchi.deleteTemplate(templateName,"","");
+            kimchi.doListTemplates();
+        },function() {});
 	});
 }
 kimchi.hideTitle = function() {
